# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegistrationInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 8 */
  phone: string;
  email: string;
  registrationUrl: string;
}

export type RegistrationStatus = typeof RegistrationStatus[keyof typeof RegistrationStatus];


export const RegistrationStatus = {
  pending: 'pending',
  paid: 'paid',
  rejected: 'rejected',
} as const;

export type RegistrationAccountStatus = typeof RegistrationAccountStatus[keyof typeof RegistrationAccountStatus];


export const RegistrationAccountStatus = {
  pending: 'pending',
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Registration {
  id: number;
  fullName: string;
  phone: string;
  email: string;
  status: RegistrationStatus;
  accountStatus: RegistrationAccountStatus;
  createdAt: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export type LoginResultStatus = typeof LoginResultStatus[keyof typeof LoginResultStatus];


export const LoginResultStatus = {
  pending: 'pending',
  active: 'active',
  inactive: 'inactive',
} as const;

export interface LoginResult {
  status: LoginResultStatus;
  message: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export type AdminOrderStatus = typeof AdminOrderStatus[keyof typeof AdminOrderStatus];


export const AdminOrderStatus = {
  pending: 'pending',
  paid: 'paid',
  rejected: 'rejected',
} as const;

export type AdminOrderAccountStatus = typeof AdminOrderAccountStatus[keyof typeof AdminOrderAccountStatus];


export const AdminOrderAccountStatus = {
  pending: 'pending',
  active: 'active',
  inactive: 'inactive',
} as const;

export interface AdminOrder {
  id: number;
  fullName: string;
  phone: string;
  email: string;
  receiptUrl: string;
  status: AdminOrderStatus;
  accountStatus: AdminOrderAccountStatus;
  /** @nullable */
  rejectionReason: string | null;
  createdAt: string;
}

export type OrderReviewInputDecision = typeof OrderReviewInputDecision[keyof typeof OrderReviewInputDecision];


export const OrderReviewInputDecision = {
  approve: 'approve',
  reject: 'reject',
} as const;

export interface OrderReviewInput {
  decision: OrderReviewInputDecision;
  reason?: string;
}

export interface AdminSummary {
  totalRegistrations: number;
  pendingOrders: number;
  approvedOrders: number;
  totalRevenue: number;
}

export interface AdminLoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  email: string;
}

export type PostInputStatus = typeof PostInputStatus[keyof typeof PostInputStatus];


export const PostInputStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface PostInput {
  /**
     * @minLength 3
     * @maxLength 180
     */
  title: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @maxLength 500 */
  excerpt: string;
  /** @minLength 1 */
  content: string;
  status: PostInputStatus;
}

export type Post = PostInput & ({
  id: number;
  slug: string;
  authorEmail: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  publishedAt: string | null;
});

export interface UploadInput {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadInput;
}

export type ListAdminOrdersParams = {
status?: ListAdminOrdersStatus;
};

export type ListAdminOrdersStatus = typeof ListAdminOrdersStatus[keyof typeof ListAdminOrdersStatus];


export const ListAdminOrdersStatus = {
  all: 'all',
  pending: 'pending',
  paid: 'paid',
  rejected: 'rejected',
} as const;