# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: registrations
    description: Course registration operations
  - name: auth
    description: Student authentication operations
  - name: admin
    description: Admin operations
  - name: posts
    description: Public news posts
paths:
  /:
    get:
      operationId: apiRoot
      tags: [health]
      summary: API service status
      description: Returns the API service status at its mounted root.
      responses:
        "200":
          description: API service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /registrations:
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Create a course registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Invalid registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginStudent
      tags: [auth]
      summary: Log in a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List course registration orders for review
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, pending, paid, rejected]
            default: all
      responses:
        "200":
          description: Registration orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminOrder"
  /admin/orders/{id}/review:
    patch:
      operationId: reviewAdminOrder
      tags: [admin]
      summary: Approve or reject a registration order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderReviewInput"
      responses:
        "200":
          description: Updated registration order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrder"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get registration summary
      responses:
        "200":
          description: Registration summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
  /posts:
    get:
      operationId: listPublishedPosts
      tags: [posts]
      summary: List published news posts
      responses:
        "200":
          description: Published posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /posts/{slug}:
    get:
      operationId: getPublishedPost
      tags: [posts]
      summary: Get a published news post
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Sign in the configured administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid admin credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth/me:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Get the current admin session
      responses:
        "200":
          description: Current admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
  /admin/auth/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: End the admin session
      responses:
        "204":
          description: Admin session ended
  /admin/posts:
    get:
      operationId: listAdminPosts
      tags: [admin]
      summary: List all news posts
      responses:
        "200":
          description: All posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createAdminPost
      tags: [admin]
      summary: Create a news post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /admin/posts/{id}:
    patch:
      operationId: updateAdminPost
      tags: [admin]
      summary: Update a news post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deleteAdminPost
      tags: [admin]
      summary: Delete a news post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted post
  /admin/uploads/request-url:
    post:
      operationId: requestAdminUploadUrl
      tags: [admin]
      summary: Request a direct image upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegistrationInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        email:
          type: string
          format: email
        registrationUrl:
          type: string
          format: uri
      required: [fullName, phone, email, registrationUrl]
    Registration:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [pending, paid, rejected]
        accountStatus:
          type: string
          enum: [pending, active, inactive]
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, phone, email, status, accountStatus, createdAt]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [email, password]
    LoginResult:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, inactive]
        message:
          type: string
      required: [status, message]
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    AdminOrder:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        receiptUrl:
          type: string
        status:
          type: string
          enum: [pending, paid, rejected]
        accountStatus:
          type: string
          enum: [pending, active, inactive]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, phone, email, receiptUrl, status, accountStatus, rejectionReason, createdAt]
    OrderReviewInput:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, reject]
        reason:
          type: string
      required: [decision]
    AdminSummary:
      type: object
      properties:
        totalRegistrations:
          type: integer
        pendingOrders:
          type: integer
        approvedOrders:
          type: integer
        totalRevenue:
          type: integer
      required: [totalRegistrations, pendingOrders, approvedOrders, totalRevenue]
    AdminLoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
    AdminSession:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    PostInput:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 180
        thumbnailUrl:
          type: ["string", "null"]
        excerpt:
          type: string
          maxLength: 500
        content:
          type: string
          minLength: 1
        status:
          type: string
          enum: [draft, published]
      required: [title, excerpt, content, status]
    Post:
      allOf:
        - $ref: "#/components/schemas/PostInput"
        - type: object
          properties:
            id:
              type: integer
            slug:
              type: string
            authorEmail:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            publishedAt:
              type: ["string", "null"]
              format: date-time
          required: [id, slug, authorEmail, createdAt, updatedAt, publishedAt]
    UploadInput:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required: [name, size, contentType]
    UploadResponse:
      type: object
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadInput"
      required: [uploadURL, objectPath, metadata]