# lib/api-zod/src/generated/types/loginResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LoginResultStatus } from './loginResultStatus';

export interface LoginResult {
  status: LoginResultStatus;
  message: string;
}