# lib/api-zod/src/generated/types/orderReviewInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderReviewInputDecision } from './orderReviewInputDecision';

export interface OrderReviewInput {
  decision: OrderReviewInputDecision;
  reason?: string;
}