# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostInputStatus } from './postInputStatus';

export interface PostInput {
  /**
     * @minLength 3
     * @maxLength 180
     */
  title: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @maxLength 500 */
  excerpt: string;
  /** @minLength 1 */
  content: string;
  status: PostInputStatus;
}