# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 8 */
  phone: string;
  email: string;
  registrationUrl: string;
}